+++ OpenSim/Common/Test/testCommonUtilities.cpp
@@ -319,7 +319,7 @@ TEST_CASE("rotateMarkerTable tests") {
         markerData(1, 0) = SimTK::Vec3(4, 5, 6);
         markerData(2, 0) = SimTK::Vec3(7, 8, 9);
 
-        OpenSim::TimeSeriesTableVec3 table(time, markerData, labels);
+        TimeSeriesTableVec3 table(time, markerData, labels);
 
         const SimTK::Vec3 marker_rotations(0, 0, 0);
         const SimTK::Rotation R =
@@ -327,7 +327,7 @@ TEST_CASE("rotateMarkerTable tests") {
                         marker_rotations[0], SimTK::XAxis, marker_rotations[1],
                         SimTK::YAxis, marker_rotations[2], SimTK::ZAxis);
 
-        OpenSim::rotateMarkerTable(table, R);
+        table.rotate(R);
 
         SimTK::Matrix_<SimTK::Vec3> resultData(3, 1); // 3 rows, 1 column
 
@@ -339,9 +339,11 @@ TEST_CASE("rotateMarkerTable tests") {
 
         REQUIRE(A.nrow() == B.nrow());
         REQUIRE(A.ncol() == B.ncol());
-        for (int i = 0; i < A.nrow(); ++i)
-            for (int j = 0; j < A.ncol(); ++j)
+        for (int i = 0; i < A.nrow(); ++i) {
+            for (int j = 0; j < A.ncol(); ++j) {
                 REQUIRE(A(i, j).isNumericallyEqual(B(i, j)));
+            }
+        }
     }
     SECTION("Valid rotation rotates the markers correctly") {
         const std::vector<double> time = {1.0, 2.0, 3.0};
@@ -353,7 +355,7 @@ TEST_CASE("rotateMarkerTable tests") {
         markerData(1, 0) = SimTK::Vec3(-4, 5, -6);
         markerData(2, 0) = SimTK::Vec3(7, 8, 9);
 
-        OpenSim::TimeSeriesTableVec3 table(time, markerData, labels);
+        TimeSeriesTableVec3 table(time, markerData, labels);
 
         const SimTK::Vec3 marker_rotations(-SimTK::Pi / 2, SimTK::Pi / 2, 0);
         const SimTK::Rotation R =
@@ -361,7 +363,7 @@ TEST_CASE("rotateMarkerTable tests") {
                         marker_rotations[0], SimTK::XAxis, marker_rotations[1],
                         SimTK::YAxis, marker_rotations[2], SimTK::ZAxis);
 
-        OpenSim::rotateMarkerTable(table, R);
+        table.rotate(R);
 
         SimTK::Matrix_<SimTK::Vec3> resultData(3, 1); // 3 rows, 1 column
 
@@ -373,8 +375,10 @@ TEST_CASE("rotateMarkerTable tests") {
 
         REQUIRE(A.nrow() == B.nrow());
         REQUIRE(A.ncol() == B.ncol());
-        for (int i = 0; i < A.nrow(); ++i)
-            for (int j = 0; j < A.ncol(); ++j)
+        for (int i = 0; i < A.nrow(); ++i) {
+            for (int j = 0; j < A.ncol(); ++j) {
                 REQUIRE(A(i, j).isNumericallyEqual(B(i, j)));
+            }
+        }
     }
 }